Screenshot tolerance 0.02 -> 0.0001: the visual gate was blind by 50x
DEFAULT_SCREENSHOT_CONFIG demanded that ~41,472 pixels of a 1920x1080 frame
differ before an assertion failed, so real changes passed green.
services/_testimonials-header sat at difference_level 0.005923 through an entire
dark-surface recolour and reported no change. Measured on
DesktopSiteTest#test_services at clean HEAD, 7 screenshots compared per run:
0.02 gives 2 failures, this commit gives 3. Run-to-run noise is 0.0 - repeat
runs return difference_levels identical to 16 decimal places, and 2 of the 7
stay green at the new floor - so 0.0001 (~207 px) clears jitter without
tripping on it. The ~30 calls that pin their own tolerance are untouched; only
the floor for calls pinning nothing moved.

Passing quietly is worse than failing, because it fossilizes the baseline: the
gem writes the fresh capture over the baseline PNG, then a PASS restores the
git-HEAD image back over it. A green run therefore never refreshes anything,
and every sub-tolerance change leaves the committed PNG showing the old render
until something large enough to fail arrives and the accumulated drift gets
accepted in one lump nobody can attribute. Corollary now documented: only a RED
run leaves a committable candidate, and deleting the PNG first does nothing
since the base is read from git HEAD.

SECTION_CONFIGS is KEPT at 0.02 on purpose. Its 7 keys equalled the old
default, making it a silent no-op; lowering the default inverts that into a
load-bearing shield holding ~22 section screenshots at the old tolerance until
their drift is measured. Deleting it here would drop those to 0.0001 unmeasured
and would also corrupt the headline number: with the shield removed the same
test fails 5 instead of 3, which is +3 rather than the +1 this change actually
costs. The two extra are exactly the screenshots whose names hit a
SECTION_CONFIGS key (services/_use-cases 0.017498, services/_technologies
0.013838), since keys match the tail after "/_" - /_technologies matches
technologies, /_testimonials-header does not match testimonials.

That mapping went from inert to safety-critical with no coverage, so it moves
to test/support/screenshot_section_config.rb and is pinned by
test/unit/screenshot_section_config_test.rb. It had to move: requiring
application_system_test_case.rb from a unit test boots Hugo, Capybara and that
file's dirty-fixtures abort. The guard was verified by mutation - splitting the
key on "-" so _testimonials-header matches the testimonials shield turns the
suite red, reverting turns it green - and the system path re-measured identical
(same 3 failures, same difference_levels) after the extraction.

test-gates.md also documents two blind spots tolerance cannot fix. Captures are
viewport-only at scroll top (1920x1080 desktop, 360x800 mobile,
test/support/setup_capybara.rb:85-86), so below-fold content is invisible at
ANY tolerance - a table-cell edit measures difference_level exactly 0 on mobile
even at tolerance 0, and every failure region here caps at the viewport height.
And perceptual_threshold = 2.0 (test/support/setup_snap_diff.rb:25) means vips
only counts a pixel as differing above CIE dE00 2.0, so a sub-threshold
recolour contributes zero differing pixels at any tolerance. Coverage fixes are
out of scope. Live doctrine swept truthful: the tutorial documented a
SCREENSHOT_TOLERANCE env var that exists nowhere in the codebase (and claimed a
1% default the code never had), a helper named
assert_stable_problematic_screenshot with its own section and three call sites,
and a per-OS tolerance-multiplier table - none of which exist in test/, bin/ or
lib/; visual-qa-criteria pinned 0.02. A doctrine file names its own mechanisms,
so its fiction looks executable and gets copied.

Gates: bin/rake test:unit 289 runs, 6090 assertions, 0 failures. bin/hugo-build
clean. okf_validate: 0 errors, 111 -> 112 warnings (exactly the one section-7
date heading this log entry adds). Screenshot suites are red on the stale #540
baselines by design - that is the gate biting:

  Screenshot does not match for 'desktop/services/_testimonials-header':
  ({"difference_level":0.005922550154320988})

which passed silently before this commit. Those baselines are NOT re-recorded
here; that is the follow-up's job.
