+++ docs/20-29-testing-qa/screenshot-testing/20.02-screenshot-testing-workflow-tutorial.md
@@ -40,7 +40,7 @@ def test_homepage_sections
   
   sections.each do |section_id|
     scroll_to find("##{section_id}")
-    assert_stable_screenshot "homepage/_#{section_id}", tolerance: 0.02
+    assert_stable_screenshot "homepage/_#{section_id}"
   end
 end
 ```
@@ -59,25 +59,6 @@ end
 def test_navigation
   visit "/"
   assert_quick_screenshot "navigation"
-end
-```
-
-### 3. `assert_stable_problematic_screenshot(name, **options)`
-
-**Use Case**: Flaky areas with dynamic content or animations  
-**Default Settings:**
-
-- Wait time: 5 seconds
-- Stability time: 2.0 seconds
-- Tolerance: 2.5%
-- Median filtering enabled (reduces noise)
-
-```ruby
-# Example usage - for dynamic content areas
-def test_blog_pagination
-  visit "/blog/"
-  scroll_to find("#pagination")
-  assert_stable_problematic_screenshot "blog/index/_pagination", skip_area: [".blog-post"]
 end
 ```
 
@@ -179,10 +160,7 @@ bin/test # Re-run to generate new baseline
 ### Key Environment Variables
 
 ```bash
-# Global tolerance (default: 1%)
-SCREENSHOT_TOLERANCE=0.02
-
-# Stability wait time (default: 1.0s)  
+# Stability wait time (default: 0.1s)
 SCREENSHOT_STABILITY_TIME=2.0
 
 # Force regeneration of all baselines
@@ -192,12 +170,6 @@ FORCE_SCREENSHOT_UPDATE=true
 CAPYBARA_SCREENSHOT_DIFF_FAIL_ON_DIFFERENCE=false
 ```
 
-### OS-Specific Tolerance Adjustments
-
-- **macOS**: 0.8x tolerance multiplier (more consistent rendering)
-- **Linux**: 1.2x tolerance multiplier (font rendering variations)  
-- **Windows**: 1.3x tolerance multiplier (highest rendering variations)
-
 ## Current Test Analysis & Recommendations
 
 ### Failure 1: `desktop/clients/_clients` (3.06% difference)
@@ -205,14 +177,12 @@ CAPYBARA_SCREENSHOT_DIFF_FAIL_ON_DIFFERENCE=false
 **Status**: ⚠️ **Review Required**
 
 - Large area affected (1.7M pixels)
-- Above problematic screenshot threshold (2.5%)
 - **Recommendation**: Manual review required, likely acceptable change
 
 ### Failure 2: `desktop/blog/index/_pagination` (2.75% difference)  
 
 **Status**: ✅ **Likely Acceptable**
 
-- Already using `assert_stable_problematic_screenshot` (2.5% tolerance)
 - Dynamic content area with blog posts
 - **Recommendation**: Review and likely accept change
 
@@ -221,8 +191,6 @@ CAPYBARA_SCREENSHOT_DIFF_FAIL_ON_DIFFERENCE=false
 **Status**: ⚠️ **Review Required**
 
 - Large area affected (2.0M pixels)
-- Above standard tolerance but below problematic threshold
-- **Recommendation**: Consider switching to `assert_stable_problematic_screenshot`
 
 ## Maintenance Best Practices
 
@@ -264,10 +232,10 @@ ls test/fixtures/screenshots/macos/desktop/
 
 ### Issue: Flaky Screenshots
 
-**Solution**: Use `assert_stable_problematic_screenshot` with higher tolerance
+**Solution**: Pin a higher tolerance on the call
 
 ```ruby
-assert_stable_problematic_screenshot "flaky_component", tolerance: 0.035
+assert_stable_screenshot "flaky_component", tolerance: 0.035
 ```
 
 ### Issue: Dynamic Content Failures
@@ -283,7 +251,6 @@ assert_stable_screenshot "page", skip_area: [".timestamp", ".dynamic-content"]
 ### Strengths ✅
 
 - Comprehensive helper method coverage
-- OS-specific tolerance handling
 - Excellent diff file generation
 - Well-organized directory structure
 - Environmental configuration flexibility
